// types/vehicle.type.ts
export type AutoPassVehicle = {
  mask: string;
  reg_country: string;
  company: string;
  entity: string;
  activities: string;
  reg_or_vin: string;
  reg_plate: string;
  vin: string;
  code: string;
  error_code: 0;
  car_identification: {
    vt_id: number;
    ktypnr: number;
    ktypnr_list: number[];
    reg_plate: string;
    vin: string;
    brand: string;
    model: string;
    version: string | null;
  };
  processing_info: {
    reg_plate_lookup: boolean;
    manufacturer_lookup: boolean;
  };
  registration_info: {
    manufacturer_id: number;
    manufacturer_name: string;
    model_id: number;
    model_name: string;
    model_base: string;
    version: string;
    generation_code: string;
    generation_code_cg: string;
    body_id: number;
    body_name: string;
    reg_body: string;
    energy_id: number;
    energy_name: string;
    gearbox_id: number;
    gearbox_name: string;
    gearbox_type: string;
    co2: number;
    depollution: string;
    date_ct: string;
    date_pme: string;
    reg_date: string;
    vin: string;
    reg_plate: string;
    type_mine: string;
    type_version: string | null;
    type_vin: string;
    manufacturing_date: string | null;
    product_key: string | null;
    vehicle_color: string;
    propulsion_type: string;
    engine_code: string;
    engine_name: string | null;
    engine_index: string | null;
    engine_size: number;
    nb_cylinders: number;
    nb_valves: number;
    turbo: string;
    injection_type: string;
    real_power: number;
    fiscal_power: number;
    kw_power: number;
    cons_ext_urban: number;
    cons_mixed: number;
    cons_urban: number | null;
    wheel_base: number;
    height: number;
    width: number;
    length: number;
    weight: number | null;
    weight_rolling: number;
    weight_empty: number;
    tire_width: number | null;
    tire_height: number | null;
    tire_diameter: number | null;
    tire_speed_index: number | null;
    serial_number: string;
    siren_number: string;
    nb_seats: number;
    nb_doors: number;
    nb_gears: number;
    nb_volumes: number;
    typevarversprf: string;
    tire_list: [
      {
        width: number;
        height: number;
        diameter: number;
        charging_index: number;
        speed_index: string;
      }
    ];
  };
};